+++ packages/document-grounding/src/client/api/retrieval-api.ts
@@ -39,18 +39,21 @@ export const RetrievalApi = {
    * List data repository by id
    * @param repositoryId - Repository ID
    * @param headerParameters - Object containing the following keys: AI-Resource-Group.
+   * @param queryParameters - Object containing the following keys: remoteName.
    * @returns The request builder, use the `execute()` method to trigger the request.
    */
   getDataRepositoryById: (
     repositoryId: string,
-    headerParameters: { 'AI-Resource-Group': string }
+    headerParameters: { 'AI-Resource-Group': string },
+    queryParameters: { remoteName?: string }
   ) =>
     new OpenApiRequestBuilder<DataRepository>(
       'get',
       '/retrieval/dataRepositories/{repositoryId}',
       {
         pathParameters: { repositoryId },
-        headerParameters
+        headerParameters,
+        queryParameters
       },
       RetrievalApi._defaultBasePath
     ),