+++ packages/document-grounding/src/client/api/schema/search-scores.ts
@@ -1,0 +1,17 @@
+/*
+ * Copyright (c) 2026 SAP SE or an SAP affiliate company. All rights reserved.
+ *
+ * This is a generated file powered by the SAP Cloud SDK for JavaScript.
+ */
+
+/**
+ * Representation of the 'SearchScores' schema.
+ */
+export type SearchScores = {
+  aggregatedScore: {
+    value?: number;
+  } & Record<string, any>;
+  denseRetrievalScore: {
+    value?: number;
+  } & Record<string, any>;
+} & Record<string, any>;