+++ packages/document-grounding/src/client/api/schema/service-now-pipeline-minimal-response.ts
@@ -1,0 +1,16 @@
+/*
+ * Copyright (c) 2026 SAP SE or an SAP affiliate company. All rights reserved.
+ *
+ * This is a generated file powered by the SAP Cloud SDK for JavaScript.
+ */
+import type { BasePipelineMinimalResponse } from './base-pipeline-minimal-response.js';
+/**
+ * Representation of the 'ServiceNowPipelineMinimalResponse' schema.
+ */
+export type ServiceNowPipelineMinimalResponse = BasePipelineMinimalResponse & {
+  type: 'ServiceNow';
+  /**
+   * @example true
+   */
+  metadata?: boolean;
+} & Record<string, any>;