+++ packages/prompt-registry/src/client/prompt-registry/prompt-templates-api.ts
@@ -77,7 +77,6 @@ export const PromptTemplatesApi = {
    * @param scenario - Path parameter.
    * @param version - Path parameter.
    * @param name - Path parameter.
-   * @param queryParameters - Object containing the following keys: includeSpec.
    * @param headerParameters - Object containing the following keys: AI-Resource-Group, AI-Resource-Group-Scope.
    * @returns The request builder, use the `execute()` method to trigger the request.
    */
@@ -96,8 +95,7 @@ export const PromptTemplatesApi = {
       '/lm/scenarios/{scenario}/promptTemplates/{name}/versions/{version}/history',
       {
         pathParameters: { scenario, version, name },
-        headerParameters,
-        queryParameters
+        headerParameters
       },
       PromptTemplatesApi._defaultBasePath
     ),