+++ packages/orchestration/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @sap-ai-sdk/orchestration
 
+## 2.11.0
+
+### Minor Changes
+
+- f72bb58: [feat] Update orchestration specification to v0.131.3
+
+### Patch Changes
+
+- cbcefb9: [fix] Route messages to `messages_history` when using an orchestration config reference or prompt template reference.
+- 7db7237: [Fixed Issue] Disallow providing both orchestration config reference and config object at the same time.
+- Updated dependencies [c8c0e41]
+- Updated dependencies [75bb9a9]
+  - @sap-ai-sdk/core@2.11.0
+  - @sap-ai-sdk/ai-api@2.11.0
+  - @sap-ai-sdk/prompt-registry@2.11.0
+
+## 2.10.0
+
+### Minor Changes
+
+- cf767a9: [feat] Update orchestration specification to 0.115.19.
+  Multiple embedding output formats are not yet supported in the orchestration embedding client.
+
+### Patch Changes
+
+- Updated dependencies [dce18df]
+- Updated dependencies [8cb466a]
+- Updated dependencies [cc9e80c]
+- Updated dependencies [cc9e80c]
+- Updated dependencies [cc9e80c]
+  - @sap-ai-sdk/prompt-registry@2.10.0
+  - @sap-ai-sdk/core@2.10.0
+  - @sap-ai-sdk/ai-api@2.10.0
+
 ## 2.9.0
 
 ### Minor Changes